pyproject.toml
[project]
name = "pythonproject3"
version = "0.1.0"
description = ""
authors = [
    {name = "Сергей"}
]
license = {text = "0.0.1"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "yfinance (>=1.2.0,<2.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "requests-cache (>=1.3.1,<2.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "black (>=26.3.1,<27.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
